simple/issuer/requirements.txt
Flask==0.12.2
PyJWT==1.5.3
cryptography==2.1.4
